Edit and update an article
Implement the edit and update methods in the ArticlesController to allow
for editing and updating an existing article in the application using
the edit_article_path (articles/:id/edit) and a form in the created
edit.html.erb view file. Refactor the form shared between the new and
edit actions into a partial that both views render.

* Add edit and update actions to ArticlesController
* Add tests for edit and update in ArticlesControllerTest
* Refactor new and edit form into _form.html.erb partial
* Refactor repeated code in ArticlesControllerTest

Completes Rails Getting Started Guide:
* 7 CRUDit Where CRUDit Is Due
  * 7.4 Updating an Article
    * 7.4.1 Using Partials to Share View Code
    * 7.4.2 Finishing Up