Use scaffold generator to create Post resource
Demonstrate how to quickly generate a resource using the Rails scaffold
generator. Generate a roughly equivalent Post method that provides
equivalent functionality to Article and modify the generated resource so
that destroy works with Turbo.

In addtion, made the following changes to Article to incorporate some of
the scaffold generator code that was better:
* Add before_action :set_article callback to refactor out repeated code
  of finding an Article
* Add display of the flash[:notice] to the Article index and show views

This was an interlude between steps 7 and 8 of the Getting Started with
Rails Guide and does not match a specific section of the guide.