Create comments for an article
Add a CommentsControler (with only a create action as-of now) and the
associated test class. Modify the articles show view to provide a list
of comments for the article and a form for adding a new comment to the
article. Modify the ArticlesControllerTest with assertions about the
new functionality for the show action and view.

Completes Rails Getting Started Guide:
* 8 Adding a Second Model
  * 8.4 Generating a Controller