Set up duplicated code for concern refactoring
Add status attribute to both Article and Comment models with the same
validation and same set of valid statuses to both models. Add test code
for new functionality to Article and Comment model and controller tests
and add new status attribute and an archived record to the fixtures for
each.

This doesn't complete any section of the Getting Started with Rails
guide, but sets things up for refactoring the repeated code into
concerns in the next commit. All tests currently passing.

In progress:
* 9.3 Using Concerns