Add CommentsMailer to notify on create and destroy
Add a CommentsMailer to the blog. The mailer notifies on creation of a
comment and destruction of a comment.

Note that because the destroyed comment is not available in the
deliver_later hook, that email needs to be called with the attributes
of the Comment instead of the ActiveRecord object.

Create a smaller partial for the comment attributes that can be called
by the mailer without running into issues with the link_to destroy.